Add nav bar 1/3: update default.html
This is 1/3 pull requests that, if merged, will give the F' site a nav bar, as shown here: https://chr1st1ansears.github.io/fprime/

This PR specifically updates default.html. It creates the nav bar and places it in a container. It renames main_content to page_content, to avoid confusion. And it adds a filler <div>, which helps with the spacing. Lastly, this PR makes the logo and tagline in the banner clickable, so users can click them to get back to the home page.